Try only setting socks proxy once
