+++ sabnzbd/getipaddress.py
@@ -80,6 +80,8 @@ def publicipv4():
     public ipv4 needs special attention, meaning forcing
     IPv4 connections, and not allowing IPv6 connections
     """
+    if sabnzbd.cfg.socks5_proxy_url():
+        return None
     public_ipv4 = None
     try:
         ipv4_found = False
