+++ sabnzbd/misc.py
@@ -25,7 +25,6 @@ import urllib.request
 import urllib.parse
 import re
 import subprocess
-import socks
 import socket
 import time
 import datetime
@@ -323,29 +322,11 @@ def set_serv_parms(service, args):
     return True
 
 
-def set_socks5_proxy():
-    if cfg.socks5_proxy_url():
-        proxy = urllib.parse.urlparse(cfg.socks5_proxy_url())
-        logging.debug("Using proxy %s:%s", proxy.hostname, proxy.port)
-        socks.set_default_proxy(
-            socks.SOCKS5,
-            proxy.hostname,
-            proxy.port,
-            True,  # use remote DNS, default
-            proxy.username,
-            proxy.password,
-        )
-        socket.socket = socks.socksocket
-    else:
-        socket.socket = sabnzbd.ORIGINAL_SOCKET
-
-
 def get_from_url(url: str) -> Optional[str]:
     """Retrieve URL and return content"""
     try:
         req = urllib.request.Request(url)
         req.add_header("User-Agent", "SABnzbd/%s" % sabnzbd.__version__)
-        set_socks5_proxy()
         with urllib.request.urlopen(req) as response:
             return ubtou(response.read())
     except:
