+++ sabnzbd/urlgrabber.py
@@ -350,7 +350,6 @@ def _build_request(url: str) -> HTTPResponse:
     if user_passwd:
         req.add_header("Authorization", "Basic " + ubtou(base64.b64encode(utob(user_passwd))).strip())
 
-    misc.set_socks5_proxy()
     return urllib.request.urlopen(req)
